+++ packages/@runlightyear/postmark/src/index.ts
@@ -1,0 +1,12 @@
+/** Postmark **/
+import { Postmark } from "./Postmark";
+import type { PostmarkProps } from "./Postmark";
+
+/** Send Single Email **/
+import type {
+  SendSingleEmailProps,
+  SendSingleEmailResponse,
+} from "./sending/sendSingleEmail";
+
+export { Postmark };
+export type { PostmarkProps, SendSingleEmailProps, SendSingleEmailResponse };